requirements.txt
streamlit==1.31.1
kafka-python==2.0.2
pandas==2.1.4
plotly==5.18.0
